Lesson 42 Init Redux
