+++ src/components/Dialogs/Dialogs.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import s from './Dialogs.module.css';
 import DialogItem from './DialogItem/DialogItem';
 import Message from './Message/Message';
-import {updateNewMessageBodyCreator, sendMessageCreator} from "../../redux/state";
+import {updateNewMessageBodyCreator, sendMessageCreator} from "../../redux/dialogs-reducer";
 
 const Dialogs = (props) => {
     let state = props.store.getState().dialogsPage;
