+++ src/index.js
@@ -1,5 +1,5 @@
 import * as serviceWorker from './serviceWorker';
-import store from './redux/state';
+import store from './redux/redux-store';
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
@@ -25,7 +25,12 @@ let rerenderEntireTree = (state) =>{
 
 rerenderEntireTree(store.getState());
 
-store.subscribe(rerenderEntireTree);
+
+
+store.subscribe(() => {
+    let state = store.getState();
+    rerenderEntireTree(state);
+});
 
 
 // If you want your app to work offline and load faster, you can change
