+++ src/redux/redux-store.js
@@ -1,0 +1,15 @@
+import {combineReducers, createStore} from "redux";
+import profileReducer from './profile-reducer';
+import dialogsReducer from './dialogs-reducer';
+import sidebarReducer from './sidebar-reducer';
+
+
+let reducers = combineReducers({
+    profilePage: profileReducer,
+    dialogsPage: dialogsReducer,
+    sidebar: sidebarReducer
+});
+
+let store = createStore(reducers);
+
+export default store;
