+++ src/redux/sidebar-reducer.js
@@ -1,0 +1,8 @@
+let initialState = {};
+
+const sidebarReducer = (state = initialState, action) => {
+
+    return state;
+};
+
+export default sidebarReducer;
